pyproject.toml
[project]
name = "floryn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.0",
    "jinja2>=3.1.6",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
]
